refactor: 도메인 패키지 그룹화
