+++ src/main/kotlin/christmas/domain/benefit/Benefit.kt
@@ -1,4 +1,6 @@
-package christmas.domain
+package christmas.domain.benefit
+
+import christmas.domain.Money
 
 class Benefit(
     private val type: BenefitType,
