+++ src/main/kotlin/christmas/domain/benefit/Benefits.kt
@@ -1,4 +1,6 @@
-package christmas.domain
+package christmas.domain.benefit
+
+import christmas.domain.Money
 
 class Benefits(private val benefits: List<Benefit>) {
