+++ src/main/kotlin/christmas/domain/order/MenuCount.kt
@@ -1,4 +1,4 @@
-package christmas.domain
+package christmas.domain.order
 
 import christmas.exception.OrdersException
