+++ src/test/kotlin/christmas/domain/DecemberEventTest.kt
@@ -1,5 +1,6 @@
 package christmas.domain
 
+import christmas.domain.order.Orders
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
