+++ src/test/kotlin/christmas/domain/order/OrdersTest.kt
@@ -1,10 +1,9 @@
-package christmas.domain
+package christmas.domain.order
 
+import christmas.domain.Money
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.ValueSource
 
 class OrdersTest {
     @Test
